+++ src/relax.rs
@@ -23,7 +23,11 @@ pub struct Spin;
 impl RelaxStrategy for Spin {
     #[inline(always)]
     fn relax() {
-        // core::hint::spin_loop();
+        // // Use the deprecated spin_loop_hint() to ensure that we don't get
+        // // a higher MSRV than we need to.
+        // #[allow(deprecated)]
+        // core::sync::atomic::spin_loop_hint();
+
         #[cfg(target_arch = "x86_64")]
         unsafe { core::arch::asm!("pause", options(nomem, nostack)); }
     }